+++ internal/RVM/core/object/event.go
@@ -31,18 +31,18 @@ type Bar struct {
 
 	FrameImageT  Transform
 	CursorSize   Vector2
-	StartPadding int
-	EndPadding   int
-	SidePadding  int
+	StartPadding float32
+	EndPadding   float32
+	SidePadding  float32
 	IsVertical   bool
 
-	MaxValue  int
-	MinValue  int
-	InitValue int
+	MaxValue  float32
+	MinValue  float32
+	InitValue float32
 
-	Down   func(e *event.EVENT_MouseButton, value int)
-	Up     func(e *event.EVENT_MouseButton, value int)
-	Scroll func(e *event.EVENT_MouseMotion, value int)
+	Down   func(e *event.EVENT_MouseButton, value float32)
+	Up     func(e *event.EVENT_MouseButton, value float32)
+	Scroll func(e *event.EVENT_MouseMotion, value float32)
 }
 
 func (b *Bar) screenObj() {}
