+++ internal/RVM/core/system/game/api_texture.go
@@ -1,6 +1,6 @@
 package game
 
-func (g *Game) ChangeTextureAlpha(name string, alpha int) {
+func (g *Game) ChangeTextureAlpha(name string, alpha uint8) {
 	g.lock.Lock()
 	defer g.lock.Unlock()
